# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de clientes",
        "title": "Customer Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/customers": {
            "get": {
                "description": "Obtiene la lista de todos los clientes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get list of customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetCustomersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Create customer",
                "parameters": [
                    {
                        "description": "Customer Data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.CustomerJson"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            }
        },
        "/customers/kpi": {
            "get": {
                "description": "Obtiene los KPIs de clientes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get KPIs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetKPIJson"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "description": "Obtiene un cliente por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get customer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetCustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un cliente existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Update customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Customer Data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.CustomerJson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un cliente",
                "tags": [
                    "customers"
                ],
                "summary": "Delete customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Verifica el estado del servicio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Simple ping para verificar conectividad",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/protected/ping": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ping protegido que requiere autenticación",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Protected ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "github_com_devpablocristo_tech-house_pkg_types.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_devpablocristo_tech-house_pkg_types.APIErrorType"
                }
            }
        },
        "github_com_devpablocristo_tech-house_pkg_types.APIErrorType": {
            "type": "string",
            "enum": [
                "NOT_FOUND",
                "CONFLICT",
                "BAD_REQUEST",
                "INTERNAL_ERROR",
                "VALIDATION_ERROR",
                "UNAUTHORIZED",
                "TIMEOUT",
                "SERVICE_UNAVAILABLE",
                "FORBIDDEN"
            ],
            "x-enum-varnames": [
                "APIErrNotFound",
                "APIErrConflict",
                "APIErrBadRequest",
                "APIErrInternal",
                "APIErrValidation",
                "APIErrUnauthorized",
                "APIErrTimeout",
                "APIErrUnavailable",
                "APIErrForbidden"
            ]
        },
        "github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.CustomerJson": {
            "type": "object",
            "required": [
                "age",
                "birth_date",
                "email",
                "last_name",
                "name"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetCustomerResponse": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.CustomerJson"
                }
            }
        },
        "github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetCustomersResponse": {
            "type": "object",
            "properties": {
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.CustomerJson"
                    }
                }
            }
        },
        "github_com_devpablocristo_tech-house_projects_customers-manager_internal_customer_adapters_inbound_transport.GetKPIJson": {
            "type": "object",
            "properties": {
                "age_std_deviation": {
                    "type": "number"
                },
                "average_age": {
                    "type": "number"
                }
            }
        }
    }
}